I20-263 : Xspress3/4 : wait for 'Capture' readback PV to update at beginning/end of scan
At beginning and end of a scan, wait for the capture readback value to change
to true/false respectively. This ensures that the file writing has started
before beginning the scan, and that the remaining frames at end
of scan are captured and written to file before file writer is stopped.
This avoids problems for fast scans and/or slow network causing frames
to be missing from beginning or end of the scan.


Change-Id: If2a7ddb178b28170b6184a36dafbc2e37385e42c