Move EclipseUtils back into uk.ac.gda.common.rcp
It contains UI dependencies. So it should be in a client-side plugin.
This is needed to allow GUI dependencies to be removed from
uk.ac.gda.common, which is a client-server shared plugin. This was
preventing GDA from starting on the RPi where SWT binaries are not
available on ARM.

Change-Id: I15aa715893126d209f30cb56c2f630bdd21a570e