Remove GUI dependencies from uk.ac.gda.common
This is a server-client shared plugin so it should not contain GUI
dependencies. They were originally introduced to enable server code to
use EclipseUtils a class providing OSGi workarounds to lookup actual
file paths in an OSGi context.

Here that version of EclipseUtils is moved back to uk.ac.gda.common.rcp
and a reduced version with no GUI dependencies is added to the package
uk.ac.gda.common.util

Change-Id: I0be0ab2318591095e77b3d89c0adf832daff2749