[DAQ-1464] move isQueuePaused() to IQueueConnection and remove isQueuePaused(String)
This method is only called externally with the name of the submission queue.
Internally it may be called from doWhilePaused for the submission queue
or status set, so only pause the queue if the queue name is the
submission queue, as there's no consumer to listen to the status set being paused.

ISubmitter now only has methods related to submitting a bean. Methods
to do with queue management have now been removed (blockingSubmit is questionable).

Change-Id: I09ebf926a7228214649ab0d258ae3cc56eae53c6