Allow server to run without default spring profiles
If default profiles are set (by setting the gda.spring.profiles.active
property), it is useful to be able to start without any profiles.

Previously, passing `--springprofiles=` without any arguments would
prevent the server starting as it would try to set the active profiles
to an array containing only an empty string.

Change-Id: Ie9bdd71e6801aefee2a7bf52463152434d7e77ba