[DAQ-1464] remove getQueue(String) from IQueueConnection and add getRunningAndCompleted
All calls to getQueue(String) either get the submission queue or the status set, so
should be calls to get these methods directly. To do this, move getStatusSet up from
IConsumer to IQueueConnection and rename to getRunningAndCompleted.
getCompleted was also renamed to getRunningAndCompleted to match.

This removes the last method from IQueueConnection that takes a queue name
rather than just using the submission/status queue names that it was configured with.

Change-Id: I83aa53f4b545fb03d58eb90198d162f64cd505ee