DAQ-1629 Make outer scannable model persistent
There is no need to save the names of the shown scannables since the en-
tire wrappers are in the mapping experiment bean which is persistent.

A side effect of this change is that after a client reset, the outer 
scannables are an opt-in rather than opt-out feature i.e. the list is
empty rather than having all the possible scannables as possible outer
axes on the UI (a good change in this commiter's opinion). 

Change-Id: If49fc912cd2afe68e012dfc72394f150c233d5b6