DAQ-1626 Handle loading of 3D maps from the queue
The recent change in the OuterScannablesSection means that it is now
possible to load a previous scan from the queue which had an outer scan-
nable not currently shown in the UI (and not in the mapping bean).

updateControls previously updated data bindings. Now it may have to re-
create the UI.

Change-Id: I7602444e8bc37a2c005c1d2a315b3c5a38a9ac08