[DAQ-1464] remove queue name from reorder/rename/replace in AbstractQueueConnection
These methods can now be removed from ConsumerImpl. Although
ConsumerImpl is now the only concrete class that extends
AbstractQueueConnection I'm choosing not to merge them into a single
class, at least for now, and instead use AbstractQueueConnection for the
queue management methods like reorder/rename/replace, and ConsumerImpl
for the code related to controlling the consumer and running the tasks
for the consumed beans.

Change-Id: I44f15d00ebbfadcfafd9821b8457e7e8bbbc73ff