I16-240 (9.9) Add ability to write measurement group to NexusDataWriter
This optionally adds the ability to write a 'measurement' group into the
NeXus file which contains the data that would be printed to the console
during a scan.

It is off by default and can be enabled by setting
gda.nexus.writeMeasurementGroup = true

If the ScanDataPoint header contains duplicate names this will fail and
log an error.

Also add debug logging to completeCollection.

Change-Id: I0da525dd9011ff93aea0147bb162b1d9a46a8bff