DAQ-1379 Basic new MScan format support
Allows running of rectangular, circular and polygonal MScans using grid
raster and spiral paths. Performs validation on the incoming command
working with existing committed classes to build a scan request and then
submit it to the event service using the same mechanism as the original
mapping_scan_commands.py script. SPEC style scanning is not supported
yet though there are foundations for this in the ScanClauseResolver.
Also, at the moment, only detectors that are Findable are supported which
is not typically true of IRunnableDevice based ones. This means that you
must have an equivalent gda.device.Detector for each one with names that
match so that Jython can identify the device.

To make use of the functionality it is necessary to import the
mscanHandler script in the Jython terminal or your localstation script
ensuring that you don't already import mapping_scan_commands.py.
Documentation on the new syntax will be available on Confluence shortly.

Change-Id: I848edb2c9d8169e7b8e563ea4de4b864dc1a40a1