[DAQ-1615] remove use of ConsumerStatus.ALIVE as this is never the case
Simplify how to decide what to show if the bean we're showing is old.
Note: I'm not sure under what circumstances we the label provider
for the ConsumerStatus column would be called with an old HeartbeatBean
as we only update the view when we get a new heartbeat bean anyway.
Nor do we show old HeartbeatBeans on startup.



Change-Id: Ida9977a80076109256c6c15b71b636175b9c0afe