[DAQ-1619] Changes to mapping-related classes for easier reuse
MappingExperimentBeanProvider:
The mapping bean in here is set to default values in Spring, but
may be overwritten from a saved configuration at run time by
(a subclass of) AbstractSectionsView.
However, only the first view to be activated should do this:
otherwise we risk overwriting changes made by another view.
This change adds a flag to MappingExperimentBeanProvider which should be
set true when the first view sets the bean. Subsequent views will then
not overwrite the bean if this flag is set.

AbstractSectionsView:
- don't cache the mapping bean: get it from MappingExperimentBeanProvider when needed.
This ensures that all subclasses use the same bean.
- add a function to return a section of the view by its class

Change-Id: I2278eb228b72dd61930528c9b7aa68c2edc1da2c