[DAQ-1464] add a mechanism to get the consumer status via the proxy
use this to get whether the queue is paused, among other properties

This works by sending a ConsumerInfo bean to the client. This includes
the consumer status - from which isQueuePaused and isActive can be
inferred. It does not include isDurable or isPauseOnStart which are
properties to do with configuring a Consumer and not required by a client

Change-Id: Iba798b0ff3ee30c7b84adca03e23353341abfae5