I18-298 Detect sample stage in ExafsSelectionHandler and inject to bean
ExafsSelectionHandler and related classes are used to create old-style
spectroscopy scans at a point selected from a new-style map. The click
coordinates are picked up and inserted into I18's SampleStageParameters
bean.

However, I18 have 2 stages ('table 1' and 'table 3'), and this workflow
does not know the difference so x, y and zName properties in the bean
were not set.

In this change, the IStageScanConfiguration is used to get the prefix
for these stages ('t1'/'t3'). If it is a valid prefix (i.e. consistent
with all axes), it will be used to populate the bean. Otherwise it'll be
ignored and a red warning label in the GUI will tell the user that axes
selection is inconsistent.

Also includes some minor refactoring.

Change-Id: I72b48bfee42c2872860d9a28c8cc255316bf866a