[DAQ-1611] Use more reliable PV to check beam mode
The TopupWatchdog previously assumed that it could check a
bean monitoring SR-CS-RING-01:MODE and that in "normal" operation
the value of this PV would be "VMX".

It turns out that this value is not guaranteed to remain the same
and is now in fact set to "DIAD".

Instead, we should check the PV SR-CS-FILL-01:TOPUPMODE, which will
have the value "Normal" in normal operation.

Corresponding changes must be made to the beans that the topup wizard
uses to access the relevant PV.

Change-Id: Ifd566a543d61453d47fc73715e077b69cdfd319d