Correct blockingSubmit method so t reads the right topic
The blockingSubmit method is only currently used by the
mapping_scan_cammands.py script which has to manually set the topic (as
opposed to status topic) name because the method itself gets the name of
the wrong topic (all other methods in the class refer to the statustopic).
This is possible in python (because it ignores Java's access modifiers)
but not in Java so it needs to change so that the new mscan syntax
implementation can use it.


Change-Id: Ia915861f0dcc8237ea5ec2cd710b7431d179232d