DAQ-1646 Add generator for single 2D point scan
SinglePointGenerator is now mapped as the generator for the existing
SinglePointModel, so that it can now be used in a scan.

PointGeneratorService will set the coordinates in the model from the
point region since these need to be the same.

The generator is backed by ScanPointGenerator's line generator instead
of the more logical array generator since the latter only supports 1D
points. The line generator is given the model's coordinates as start and
end points which results in the desired single point.

Change-Id: Ibb54ffbac9510dea83f08f9f15bc10623c089c24