DAQ-1661 - Restore icons lost by move to Oxygen TP
The issue was we are reusing built in eclipse icons for the project
explorer and the problems view. Unfortunately Eclipse decided to change
icon format from png to gif and we had hard coded the paths. Searching
turned up this
https://www.eclipse.org/forums/index.php/t/1088678/
which helped me to find the issue.

Also adds a null check so in the future this issue would cause the
client to fail at startup.

Change-Id: I8cafba85f935bdbfdc171652df17dac0bd73bc8e