Replace use of "All" Test Suite with test discovery
If you use a single Test Suite, then when tests are run in Jenkins, it's
impossible to see which individual test is currently running. This makes
trouble-shooting problematic. This changes fixes that.

This change selects the same tests to run as those specified in the
"All" Test Suite (which itself is a subset of the total test classes),
with a couple of small differences:

(a) The original Test Suites included this test twice:
        org.eclipse.scanning.test.scan.preprocess.PreprocessTest
    The new releng.ant only runs it once

(b) The original Test Suites did not include this test:
        org.eclipse.scanning.test.scan.AnnotationScanTest

Change-Id: I4e5c7cb3b7a5408b39fd77263d82b9e6f3830d33