[DAQ-1464] use CountdownLatch instead of semaphore to avoid race condition
In QueueManagementTest.startConsumer(), if the InitialProcess was
started and tried to acquired the semaphore again (just after releasing
it) before the main thread in startConsumer tried to acquire the semaphore,
the test could hang with a deadlock. This has been fixed by using
two CountdownLatches, one for each situation that we need to wait for.

Change-Id: I74ddf61bb14856271b9a33a5b4355f55570a2491