DAQ-1567 Get template regions from IMappingRegionManager
Every call to getTemplateRegions() (previously getRegions()) now returns
a *copy* of the configured region instances so that the default/template
parameters configured in Spring are not lost.

Additionally, a new method getTemplateRegion(Class) will return the sin-
gle template of the given type.

To achieve this IMappingScanRegionShape now has the method copy() which
returns a field-by-field copy of the instance.

Change-Id: I1d1adca86652e5c47fb3f0abdede4e475bdd90ff