[DAQ-1464] fix moving up/down the queue. The indices were the wrong way around
Note that the StatusQueueView shows the queue 'upside down' in that
the tail of the queue is at the top and the head of the queue is at the
bottom.
The ConsumerImpl.reorder method reverses the queue before re-adding the
bean at the new index, meaning that passing an amount of 1 means moving
the item towards the head of the queue, and an amount of -1 moves the
item towards the tail of the queue.

Change-Id: Icb983d3476d2156fb6f62df2f362552f99f895d2