Restore the ThresholdFilter on the LogPanel appender as a placeholder
Compared to setting the root logger level, using a ThresholdFilter allows
the LogFile to continue receiving all messages while applying a filter just
on messages going to the LogPanels. This is useful if you need to enable
a noisy TRACE logger, but don't want those messages flooding the LogPanel.

Adding the ThresholdFilter to the appender also allows this level to be
changed easily while still using the standard logging configuration.

Also update and clarify the out of date root logger documentation.

Change-Id: Iba03597e67ef1f292ed259016b182d72f320ab01