DAQ-1707 Allow Malcolm's run method to run for longer than 1 second
RPCClientImpl::request takes a timeout which applies to both creating a
channel and the actual request (e.g. Malcolm::run) returning. The time-
out is changed here from 1 second to 2 days.

However we do not want to wait 2 days to find out that we cannot make
the connection. For this reason this change also creates a test channel
with a timeout of 1 second and immediately destroys it.

Change-Id: I195515add299ee2ff66a3a784a961506f7064f72