DAQ-1678 Make DataPointCache findable
Caches are currently set up in localstation and therefore cannot be
reused across namespace resets. This leads to multiple unreachable
caches that all hold scan data for each scan.

This change allows caches to be created and configured in Spring and
reused when setting up post scan prpcessing (the main use of the cache).

Change-Id: I328a8c64544a0a200a3b2ed1cd35d69208a93e39