DAQ-1739 : Replace Mucal with XrayLib
Changes to replace Mucal with calls to XrayLib :
- New function added to PressureCalculation to calculate absorption coefficient
using XrayLib functions.
- Default is to use XrayLib absorption coefficient values during pressure calculation;
the Mucal part can be removed completely once everyone is happy it is working ok.
- Some tidying up in call to PressureCalculation in IonChamberComposite,
since no longer necessary to run it as a separate 'task'.
- New test added to generate absorption coefficients using Mucal and XrayLib.
This generates series of text files for plotting to validate the results.
 It it intended to be run manually when needed rather than as part of Jenkins unit tests.
- Rebuilt version of Mucal for RHEL7; this is not backwards compatible with
RHEL6 due to changes in system libraries. The binary is included in the
unit test package.

Change-Id: I059f07a6ea7aabdb9b495e7828dd13ddb7e2ea1e