DAQ-1751 Let RmiAutomatedExported work with RBAC
getClass for RBAC wrapped objects returns cglib proxy Factory for all
instances so a ServiceInterface annotation used by the RMI exporter
could not be found.

This uses the same method as the previous CORBA implementation to check
if objects are wrapped and if so to check the type of the underlying
object instead.

Change-Id: I9ff8636f29813e2d0464c99ff380342c13fd83b2