Add logging of long running RMI-RPC calls from the UI thread
Generally it is considered bad practice to make RPC calls from the UI
thread, as it will freeze the UI. This change will help find when and
where these calls are being made.

The detection of the UI thread relies on the fact that calls from the
UI thread will be on a thread with the name "main". This mechanism is
fragile, but means that we don't have to add SWT dependencies to core.

Long running RPC calls from non UI threads are also logged at debug
level, as these are necessary in many circumstances, but may indicate
secondary effects of a slow UI.

This logging should help in general debugging of slow event handling.

Change-Id: I3e540ad6935da485812271ef3658a94395e2a3f0