DATAACQTEAM-192 Use Spring bundle class loader when creating RBAC proxies
This should be reliable, if Spring can instantiate the class then it
must be able to load it. Therefore using the same classloader again when
RBAC wrapping it makes sense. The same approach is used for RMI
exporting.

This should allow the remaining buddying on uk.ac.gda.libs to be removed
and should remove the need for "fake" package imports on net.sf.cglib.*

Change-Id: Iaced4e2b96b5947ebfff1fa1567a3186b76e9815