Wrap the log server SocketAppender with an AsyncAppender
This helps to ensure logging never reduces client or server performance
and also ensures that threadName is always populated.

Without this change, log messages with a null threadName end up with
the log server filling the thread name as "logback*".

This change follows on from a recent change to graylog_appender.xml for
DAQ-1679.

Change-Id: I052e81f0561920041bdeb792bbeba58d6e5c1e5c