DAQ-1780 Make AcquireRequestHandler only throw real EventExceptions
The design of Requester/RequestHandler seems to be that if the handler
fails for some reason, it should set the bean status as FAILED with its
message taken from the exception and then returned to the requester.

Given that the signature specifies that process() throws EventException,
and that when we throw we never return the bean, it may be reasonable
to conclude that mangling all exceptions into EventExceptions and re-
throwing breaks	the original design.

With this change:
 - Exceptions are not repackaged into other types
 - process(...) catches and immediately throws EventException
 - process(...) also catches Exception, sets the bean status to FAILED,
 	sets the bean's message as the exception's root cause message, logs
 	the exception, then returns the bean.

Change-Id: I410c64db6589d0f3668039146df5b49a2db06062