[DAQ-1709] remove lastAlive and regenerate hashCode and equals methods of HeartbeatBean
also remove equalsIgnoreLastAlive. Note that the publishTime field is
not ignored by the equals methods. ConsumerView.sync previously
calls equalsIgnoreLastAlive, but it seems this would not have updated
the UI when a new bean was received as the beans would have been equal
due to lastAlive and publishTime not being part of the comparison, so
the lastAlive column of the view's table would not have been updated

Change-Id: I4449311fbde1da86fb6c655f5d467576ac94bbab