DAQ-1331 Improve successful startup reporting
In the old logic successful startup would be reported unless specific
failure conditions were returned. This has been changed so that failure
is the default unless success is specifically returned.

Also added a case to give a warning if the presence of the startup file
indicates the server has started but the status ping does not respond
within 60 seconds after that, e.g. if the port has been set incorrectly
for some reason. This timeout will only come into play once the startup
file has been written which is the last thing the GDA application does
before returning from startup; if the ping gets a response before the
file is present, successful startup will still be detected.

Change-Id: I3b302f7b30aa0524cc1cc79b5bd2ce27c650a88b