DAQ-1799 : Make tab order in 'Experiment' view configurable
The scan tab order used in Experiment perspective is now configurable
by setting values in plugin_customization.ini. New settings
exafs.editor.scan.tab.order can be used to set the order of the tabs and
exafs.editor.selected.scan.tab can be used to set the type of the default
selected tab. Default order is the same as before if no values are given.

Change-Id: If67a5d6aaee391e31c4fa6c2eb57b72696af2e75