DAQ-1719 : Generalised version of detector rate view(s)
- Server side DetectorMonitorDataProvider object has been rewritten. It now
has a general 'collectData' function that collects a single frame of data for
supplied list of scannables and returns a list of the readout values
(as would be produced on Jython console during a step scan).
- A client side DetectorRateViewConfig Factory bean should be created
in spring for each detector rates view required. This contains the list
of detectors/scannables to collect data from in the view.
- Each client side detector rate view is created from a DetectorRateConfig object.
This has controls to start/stop the collection and shows a table with the
latest detector data. The number and names of the columns in the table
are set using getExtraNames and getName function of the scannables used
during data collection.
- DetectorRateViewFactory can be be used to generate a view in the 'view'
extension point of plugin.xml by adding the name of the DetectorRateConfig
object to the class name.
- Old detector rate views have been removed

Change-Id: I06ac62dd47d9467d9176ccf642507826a64a47f0