EpicsMonitor - Improve performance by removing unnecessary calculations
Previously every time an update was received the percentage change was
calculated requiring a x100 then compared with the sensitivity. Now the
sensitivity is stored internally as decimal moving the scaling factor to
only when the sensitivity is get or set instead of every update.

For arrays rearrange the equation to remove an unneeded array lookup.

For arrays return once an update event is sent to avoid multiple updates
when many array elements are changed.

Change-Id: I21f8d532ae68f134c6a249b29246e807413b754f