DAQ-1719 : Xspress4 updates for detector rate view collection
- Xspress4 - atScanStart now only sets up the hdf writer plugin PVs if 
HDF writing is being used for the scan; If atScanStart is called outside
of a scan, number of frames to collect is set to 1.
- To avoid confusion and the possibility of inconsistent parameters, 
Xspress2CurrentSettings.is/setAddDTScalerValuesToAscii() now delegates
to XspressParameters.set/isShowDTRawValues(). This means that calls to Xspress2CurrentSettings.setShowDTRawValues(...) propagate down to the
XspressParameters object and correctly update the output format.




Change-Id: I30a923b67104142c7aa79731170c55e7ebb45f28