[DAQ-1756] Move NumberAndUnitsComposite-related functions out of UIHelper
The particular parameters with which they create the composite
(such as mm for model units) are specific to Mapping.

The functions have been moved closer to the places where they are
used (AbstractModelEditor & FocusScanUtils).

Note that the code to handle the uk.ac.gda.client.defaultUnits
property will soon be deleted, as the initial units will be
set in Spring - see I14-550.

Change-Id: I0f116a809f303a26687d02f3e77bf173b0632536