Stop updating state in FindableProcessorQueue.update for non-pause events
Let the run method and co process the relevant command and update
state accordingly, allowing a command to be submitted as soon as an
observer receives the notification.

Pause/Unpause must still be handled in update because the command's
run method cannot be assumed to terminate, preventing the run loop
from processing the command.

Change-Id: I9c9709c96f72bcd949f679c3234145e74d454947