Remove autoScaleOnAcquire button since it duplicates Rescale action
If checked, the checkbox in FluoDetectorAcquireComposite overrides the
action to 'Rescale axis when plotted data changes' in the plotting
system (confusing). The label suggests that rescaling would happen on
acquisition only, but in fact it occurs when changing the selected ele-
ment. Effectively the logic of these two functions is the same, so this
change removes the button in the composite in favour of the standard
plotting system one; the plot behaviour is set initially to rescale when
the data changes.

Change-Id: I8e330fb7ed22251a69eb2e11bf7377de8ad41e84