Remove unused RegionJob
It only has one constructor which is never called. The only method which
takes a RegionJob in SequenceView is never called so it is also removed.
As the Job is removed the need to schedule it is gone so RegionJobRule
is also removed.

The package org.opengda.detector.electronanalyser.client.jobs which
contained these classes was not exported so there can be no users from
Jython.

Change-Id: Ic08520613e43063e76736c0b650f527c0b883154