I20-393 : Make multiscan xml file selection work correctly.
This change fixes a bug with changing xml files in the multi-scan view
using the dropdown combo box where scans do not get updated with with
the newly selected file.
This was caused by the modify event (run after combo box selection finishes)
getting passed a null scan object and so the scan object and table were
not updated to reflect the user change.
This changes fixes the problem by getting the scan object currently
being modified from the selected row in the table.

Change-Id: I299b327bcc3a49733d5ff1fcb14ea5296b6bc7e0