DAQ-1777 Fix case where completions could be lost
If there were two possible completions with the same text e.g.
completing on a Java class with a public abc field and a abc() method,
then the completion would be dropped by the TreeSet as the compareTo
method returned 0. Now in that case the type will also be considered.

Add hashcode and equals with the same logic as recommended when
implementing compareTo.

Change-Id: I131d7119c92b1df6e7dd57dead5decea0e707462