[DAQ-1823] Refactor ScanPathEditor out into a separate class
The scan path text field and associated button to display an
editor has been moved out of OuterScannablesSection, so that it can
be used elsewhere.

As it now communicates with OuterScannablesSection via a listener,
which must be deleted when the GUI is closed, dispose[Internal]()
functions have been added to AbstractMappingSection and
MappingExperimentView to propagate disposal down to ScanPathEditor.

Change-Id: I39916c3d851a9d4977b1097bd45323c4da862ac0