Refactors UpdatePlotQueue to use a BlockingQueue
This means the thread can now wait on items to come into the queue
instead of polling the queue and then copying all the items to an array.

This make the performance of the Scan Plot much better. It is now much
more responsive to showing/hiding traces which happens instantly,
instead of after a up to 500 ms wait.


Change-Id: Ia15db180a25551a0b9a916be3fada115247b6967