Change default processing timeout to 10 minutes
This timeout specifies the time allowed for the processing consumer to
wait for the unique keys to update. The previous default of 1 minute is
problematic when doing a snake scan and a single line takes over 1 min;
the processing consumer does not iterate snake-wise, so on reverse lines
will always timeout waiting for the data to arrive.

Change-Id: Id5e7a07f686be86f35f3bf7109e4efd8b84cb982