DAQ-1861 Add option to create positioners group linked to before_scan
This add a property 'gda.nexus.linkPositionersGroup' which if enabled
will create a link called 'positioners' to the 'before_scan' group which
typically contains the "postion of scannables before the scan".

The use of 'positioners' is used by ESRF so some software e.g. pyMCA
expects it to be there.

Change-Id: I76b3b35dde7f9ffdd2b48e6c909138bcb4ccb138