DAQ-1902 Don't give Async threads the name of the calling thread
If a thread name included formatting characters, Async would try and use
it as a format string and throw an exception. It could be possible to
filter out formatting characters but giving new threads the same name as
previous threads is a source of confusion.

If a name is required, one can be given explicitly.

Change-Id: Idc33e87a20de22a96088a6c2289c68793cbb33e5