DAQ-1612 Fix invisible radio buttons in Close Menu dialog (UserOptionsMenuOnClose)
The main change is to add a 2×4 grid (buttonComposite) and have bare radio
buttons in the first column, and labels (that wrap) in the second column.

Because the radio buttons themselves no longer have the SWT.WRAP bit set, they
aren't invisible when the dialog first opens.

I've also made a number of other tweaks to the layouts / layout data to make
the controls behave more sensibly when the dialog is resized.

Change-Id: I99ce231812bf772e59a318d837e40f778013c1ad