DAQ-1916 Add a scan submitter for the IPlan supporting two priorities
The purpose of this class is to prevent a queue of submitted scans and
keep some level of synchronisation between an event which would cause
the submission of the scan and the actual execution of said scan.

Submission is trivial when the queue is empty. For non-empty queues, we
prioritise the running scan and throw ScanningException on the submis-
sion of a new one, unless we deem the new scan to be 'important'. In
which case, this submitter will attempt to terminate and clear running
and submitted jobs to make space for the new important scan.

Change-Id: I51a56d691767a224058ce65a7e7b2b54746d85dc