DAQ-1934 Improve reliability of TriggerTest
timedTriggerShouldStopWhenDisabled no longer tries to predict the time
the trigger was active for, since that is not the point of the test.

timedTriggerHasReasonableDuration marked as @Ignore: mocking the timing
element of this test would defeat its purpose, but for reliability the
trigger duration would have to be significantly longer (too long for CI
unit tests).

Per-event timeout in countdown latch reduced, but now using the return
value of await().

Change-Id: Id336631b9724bd76237de498797aefbd4f0cf9e9