[DAQ-993] Allow initial units to be specified individually for path specifications
Currently, there is a single setting for the initial units to be shown
for each field in a path specification (xStart, yStart etc.), which
is set as a Java property uk.ac.gda.client.defaultUnits (defaulting
to millimetres if this property is not set).

This is not ideal for some beamlines: for example on I14, it would
be most natural to specify a starting point in mm but a step size in nm.

This change allows you to define an InitialLengthUnits bean, which
must be registered as an OSGi service and wraps a map of field name
to initial unit. If this bean/service does not exist, GDA defaults to
the old behaviour.

The field names in the map are those already used for binding each
field and are defined in AbstractModelEditor. An example of defining
the bean is in mapping.xml for I14.

Change-Id: I8410bfeb4ef5b716bdc0109fb82f4a359ca07d7f