DAQ-1928 Revert EPICSv4 jars to use older versions
This removes the version specifiers from the epics jars, so that whatever
version is in the p2 site for the tp/ is used.

It also removes the dependency on additional jars that were added back when
we moved to the newer version, but are not required now that we are moving
back to the earlier version.

This is a partial revert of Ie4cd498bea745b94e065e5806d1b41be7323e6c3

Change-Id: I301e69d99ea3cee71261d246115dca4d3cb3dcf3