Fix concurrency issues in segments' and triggers' event-handling methods
SegmentBase.terminateSignal is now synchronised and uses 'activated' to
prevent consecutive events terminating the segment multiple times.

TriggerBase.signalChanged is now synchronised and uses 'evaluating' for
similar reasons, which is toggled in a try/finally block.

Some implementations of TriggerBase.evaluateTriggerCondition have side-
effects at the moment which is why the evaluation of the event is also
synchronised for TriggerBase, unlike in SegmentBase. To be fixed later.

Change-Id: I8e6be430b1a66a678621f221ef2cb958cf840048