[DAQ-1924] add PREPARING and FINISHING statuses, set by ScanProcess
PREPARING is the status the ScanBean is set to when ScanProcess is
created and maintained during running a before script and/or moving to a
start position, if either of those actions are performed

FINISHING is the status the ScanBean is set after the scan itself
is finished and maintained while running an after script and/or moving
to an end position, if either of those actions are performed

The status PREPARING has replaced the previous status QUEUED which was
misleadingly named as scans only had the status after they had been
taken off the queue, until they started running.

ScanProcess is now responsible for setting the status of a scan to
COMPLETED. AcqusitionDevice.execute should leave the ScanBean
with the status of RUNNING (assuming it completed normally).

ScanProcess now also has the responsibility of checking before each
stage of the overall scan (running before/after scripts, moving to
start/end position and the actual scan) whether abort has been called,
and if so throwing an InterruptedException instead of continuing to
perform the scan. Additional test methods have been added to
canProcessTest to check that this is done.

Additionally some code clean-up of ScanProcess, including not passing
around the ScanRequest, which can be obtained from the ScanBean,
moving the preparation tasks to the start of the execute method and
adding comments.

Change-Id: I6e62ea862662e607d8affeab8061ea71c87f872d