Don't treat NameError as special case in JythonServer
When evaluateCommand throws an exception, it is only logged if the
exception doesn't contain NameError. This doesn't help with tracing
problems and hiding info from the user (as the comment said) is less
useful than being able to debug things.

PyExceptions are also converted to a form useful for logging. Error
with new format:

    2019-01-29 13:49:59,420 ERROR gda.jython.GDAJythonInterpreter - Error evaluating command: abcd + 17 [2019-01-29T13:49:59 : abcd + 17]
    gda.jython.logging.PythonException: NameError: name 'abcd' is not defined
        at <string>.(script)(<module>:1)

Change-Id: I64ee39e8e20486b481f6797bbd76d4595700896e