DAQ-1945 Prevent infinite loop trying to resolve GdaRmiProxy
In the case where no RmiProxyFactory was present in the context a loop
was created by trying to resolve GdaRmiProxys in order to determine
their type, and therefore if they could be a RmiProxyFactory. This would
never be the case. The fix here is to search the context but disallow
non-singletons and eager initialisation which attempt to resolve
FactoryBeans to their objects.

Adds a test which demonstrates the issue which passes with this fix.

Also some minor improvements to the docs and exceptions.

Change-Id: I0aec9d0c15f24532a109e0b3ed68b0cd2b6989c3