I20-404 : Improvements to ADRoiCountsI0 used for medipix FFI0
Improvements to NXPlugin used to generate medipix FFI0 values :
 - Throw exception and show warning on Jython console if roi1_total counts
could not be found for detector (this is normally due to the ROI not being set).
 - If i0_channel is set to -1, the first non-time readout value from counter
timer is used for the I0 counts. I0 is always first value returned after
optional time frame value so this change ensures correct I0 value is
used whether time is included in readout or not.
 - Some other minor code cleanup was also done.

Change-Id: I7c3db2a1637a9fafb3c645f4564e2c0a73b60c7a