I20 : Bugfix for XasAsciiNexusDataWriter when doing multi-dimensional scans
This fixes a bug recently introduced caused by creating the NexusDataWriter
in 'configureScanNumber' in order to have the filename set early in
the scan (commit 2178321ab5). This function is called at start of each
inner loop of a multi-dimensional scan, creating a new NexusDataWriter each
time and causing the scan to fail.

Change-Id: Ic7f8e3a3e84b847c8e38bacece9ad088c3420131