[DAQ-2005] remove IDeviceDependentIterable and add getNames to point generator
IDeviceDependentIterable was written before the methods getShape()
and getRank() were added to IPointGenerator and the only way to get
these was to generate the points first. These methods do now exist
so the point generator may not need to generate the points. Besides
which IDeviceDependentIterable is never used outside of tests.

The method IPointGenerator.getNames() was added to reduce the duplication
of code to get the names from the first point, and so that if in future
we may be able to get the names from the jython scan point generator,
we only need to change the code in one place.

Change-Id: I0ffb02ac07519e8899bf9c487fbb043fbc89563e