I20-413 : Make ADRoiCountsI0 work with any ROI name
This change removes the need to have the detector ROI name set to
'roi1' in order for the ROI counts to be looked up correctly.
The name of the ROI to be used is now determined at the start of each
scan by examining the ROI contained in the ADRoiStatsPair plugin of
the NXDetector.

Change-Id: Ie8b7981aa76434a285f8d7af926297225cc267b6