I20-409 : Implement analyser selection for XES spectrometer scannable
This change makes it possible to choose which of the XES analyser
crystals are moved when changing the spectrometer energy.
- The scannables associated with the 3 crystal analysers and the detector
are now stored in their own scannable groups.
- Each scannable group has an 'allowedToMove' flag. If this is set to false,
none of the scannables in the group will be moved by calls to 'asynchronousMoveTo',
'isBusy()' returns false and 'stop()' has no effect. This should allow an analyser
with hardware faults to be deselected and completely ignored when changing
spectrometer energy.
- The additionalCrystalsInUse array is now used to select which analysers
are to be moved when changing energy.
- The trajectory position calculation and move have been refactored from
'asynchronousMoveTo' into separate functions.
- Unit tests have been added to check positions of detector and analysers
are correct, and to make sure analysers only move when they have been selected.

Change-Id: Ie487086393fe2cee5ac2b098aa6d29768d1c8105