I23-712 Allow property to force stop on all Stoppables via Console stop
The 'stop' action on JythonTerminalView differs from the StopAll button
in that the handler does not call stop() on all Stoppables - the
existing property (gda.jython.hardware.stop.immediately) can force stops
on Scannable and Detector interface implementers only.

I19 and I23 use Stoppable for non-Scannable devices to ensure reset
operations are executes on a halt.

To get the same behaviour from both stop mechanisms, the options were to
(a) override the handler just for I23, (b) make the existing property
handle all 'Stoppables', or (c) use an additional property for the
desired behaviour on the existing handler

This change implements option (c) so as to avoid changing default
behaviour on existing beamlines, and leave the option explicitly opt-in
for all beamlines.

Change-Id: Ibf318a5aba5453e1dfdbe129011e3a207042aeb8