DAQ-1992 Add live control implementation for enum positioners
A base class (AbstractPositionerComposite) has been extracted from
NudgePositionerComposite to handle the update readback job as well as
the label and stop button. Positions in this class are typed as Object.

ScannablePositionerControl now creates the standard composite
NudgePositionerComposite unless the scannable is an EnumPositioner, in
which case it creates an EnumPositionerComposite. This new composite
simply adds a combo box showing the positioners enumerated positions.

Change-Id: Icc4907bb1f088f436ae123f93b799888e3037be8