[DAQ-1840] Make LinearFunctionTest independent of toString()
- Currently, LinearFunction takes quantities in String format as input.
This will soon change, but for the time being, modify the test so
that the input strings are set in constants and the Quantity objects
are created from them, rather than vice versa.

This makes the test independent of Quantity.toString() (which in any
case is different in jscience4 and will cause the test to fail.

Change-Id: I62030e3e15584a2dbf93bc83f51557786ddc2cbf