DAQ 1956/2023 Add Region and Path Controller
The RegionAndPathController is intended to act as the controller (in
and MVC) way for views that need to intigate and be aware of changes
to the mapping experiment bean. The RegionAndPathSection is modified to use this and it is added to the common mapping config so that it is available for all such views.

Also move plotting handlers to controller from MappingExperimentView
registering with Eclipse context so that controller can respond to
updates posted by PathInfoCalculationJob.

Change-Id: I121a8bd890215b4d2132928f2ea7f3fdd2904d17