Set ConverterNameProvider in AutoRenameableConverter constructor
Used to pass the provider name, then used the Finder to find the actual
provider. For some time this has produced a warning at server startup
since the provider object had not been created yet.

Since both this and the ConverterNameProvider are created by Spring we
can pass in the actual object which fixes this.

Also deleted commented out code and useless javadocs.

Change-Id: Ic97c804a92747db6e7bab9d03dcdcc5747bf34db